+++ SolutionInfo.cs
@@ -3,11 +3,11 @@ using System.Runtime.InteropServices;
 
 [assembly: AssemblyCompany("Aaron Powell")]
 [assembly: AssemblyProduct("Chauffeur.ContentImport")]
-[assembly: AssemblyCopyright("Copyright © 2014 - 2015")]
+[assembly: AssemblyCopyright("Copyright ï¿½ 2014 - 2018")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
-[assembly: AssemblyVersion("0.1.0.0")]
-[assembly: AssemblyFileVersion("0.1.0.0")]
+[assembly: AssemblyVersion("0.2.0.0")]
+[assembly: AssemblyFileVersion("0.2.0.0")]
 
 [assembly: ComVisible(false)]