+++ releasenotes.md
@@ -1,3 +1,8 @@
+## New in 0.2.0 (Released 2018/01/03)
+
+* Updated to the latest Chauffeur stable build
+* Added a new Deliverable `content-publish` to allow you to publish content after importing
+
 ## New in 0.1.0 (Released 04/09/2015)
 
 Initial release of the plugin